requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
jinja2==3.1.5
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
pydantic==2.10.2
pydantic-core==2.27.1
pydantic-settings==2.6.1
pygments==2.18.0
pyjwt==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
pyyaml==6.0.2
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typer==0.14.0
typing-extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
